+++ ydb/library/yql/tests/sql/dq_file/part10/canondata/result.json
@@ -1147,9 +1147,9 @@
     ],
     "test.test[in-in_sorted_by_tuple--Debug]": [
         {
-            "checksum": "25eac6930d6256ef93f7c557aa32f2ce",
-            "size": 6188,
-            "uri": "https://{canondata_backend}/1931696/3b66a4b8ee9789607df97fd1b710d3ca890dd9b2/resource.tar.gz#test.test_in-in_sorted_by_tuple--Debug_/opt.yql_patched"
+            "checksum": "95aaa80a7fec696a32886d16fb9379c7",
+            "size": 6169,
+            "uri": "https://{canondata_backend}/1946324/4e55210d05969637e5668c50c7fbdc1a61f108f5/resource.tar.gz#test.test_in-in_sorted_by_tuple--Debug_/opt.yql_patched"
         }
     ],
     "test.test[in-in_sorted_by_tuple--Plan]": [
