+++ ydb/library/yql/tests/sql/dq_file/part18/canondata/result.json
@@ -1543,23 +1543,23 @@
     ],
     "test.test[join-premap_map_cross--Analyze]": [
         {
-            "checksum": "abb0f979cc72a099d377285fab8c7bed",
-            "size": 7402,
-            "uri": "https://{canondata_backend}/1880306/e4ccac619cc79d4b07e7e803e386d47da238c793/resource.tar.gz#test.test_join-premap_map_cross--Analyze_/plan.txt"
+            "checksum": "b53c2b9a3afadf54cfd6f457910a138b",
+            "size": 7416,
+            "uri": "https://{canondata_backend}/1130705/ac9859ee8d53f34b0483c0f88da3629c3f1f4324/resource.tar.gz#test.test_join-premap_map_cross--Analyze_/plan.txt"
         }
     ],
     "test.test[join-premap_map_cross--Debug]": [
         {
-            "checksum": "ab59bf52e40d4db6adfc196336634397",
-            "size": 3296,
-            "uri": "https://{canondata_backend}/1946324/52216c80dd82a135143a59b93de6ed80422b3740/resource.tar.gz#test.test_join-premap_map_cross--Debug_/opt.yql_patched"
+            "checksum": "5393179fa0e3ad34e02ef23a3367a644",
+            "size": 3281,
+            "uri": "https://{canondata_backend}/1130705/ac9859ee8d53f34b0483c0f88da3629c3f1f4324/resource.tar.gz#test.test_join-premap_map_cross--Debug_/opt.yql_patched"
         }
     ],
     "test.test[join-premap_map_cross--Plan]": [
         {
-            "checksum": "abb0f979cc72a099d377285fab8c7bed",
-            "size": 7402,
-            "uri": "https://{canondata_backend}/1880306/e4ccac619cc79d4b07e7e803e386d47da238c793/resource.tar.gz#test.test_join-premap_map_cross--Plan_/plan.txt"
+            "checksum": "b53c2b9a3afadf54cfd6f457910a138b",
+            "size": 7416,
+            "uri": "https://{canondata_backend}/1130705/ac9859ee8d53f34b0483c0f88da3629c3f1f4324/resource.tar.gz#test.test_join-premap_map_cross--Plan_/plan.txt"
         }
     ],
     "test.test[join-premap_map_cross--Results]": [],
@@ -2925,23 +2925,23 @@
     "test.test[tpch-q15-default.txt-Results]": [],
     "test.test[tpch-q19-default.txt-Analyze]": [
         {
-            "checksum": "44fd44c0268c4aa0aa8b5a89b9171a12",
-            "size": 8703,
-            "uri": "https://{canondata_backend}/1880306/e4ccac619cc79d4b07e7e803e386d47da238c793/resource.tar.gz#test.test_tpch-q19-default.txt-Analyze_/plan.txt"
+            "checksum": "ddd5883059984a329a12a44b3ad1bfdc",
+            "size": 8710,
+            "uri": "https://{canondata_backend}/1130705/ac9859ee8d53f34b0483c0f88da3629c3f1f4324/resource.tar.gz#test.test_tpch-q19-default.txt-Analyze_/plan.txt"
         }
     ],
     "test.test[tpch-q19-default.txt-Debug]": [
         {
-            "checksum": "3032ad5719d8b6e90ee2e441848a2df4",
-            "size": 7223,
-            "uri": "https://{canondata_backend}/1773845/57222273edb3e600187ff62653acf03a13f24744/resource.tar.gz#test.test_tpch-q19-default.txt-Debug_/opt.yql_patched"
+            "checksum": "7775ab18865a75acb0011c8a047055fb",
+            "size": 7212,
+            "uri": "https://{canondata_backend}/1130705/ac9859ee8d53f34b0483c0f88da3629c3f1f4324/resource.tar.gz#test.test_tpch-q19-default.txt-Debug_/opt.yql_patched"
         }
     ],
     "test.test[tpch-q19-default.txt-Plan]": [
         {
-            "checksum": "44fd44c0268c4aa0aa8b5a89b9171a12",
-            "size": 8703,
-            "uri": "https://{canondata_backend}/1880306/e4ccac619cc79d4b07e7e803e386d47da238c793/resource.tar.gz#test.test_tpch-q19-default.txt-Plan_/plan.txt"
+            "checksum": "ddd5883059984a329a12a44b3ad1bfdc",
+            "size": 8710,
+            "uri": "https://{canondata_backend}/1130705/ac9859ee8d53f34b0483c0f88da3629c3f1f4324/resource.tar.gz#test.test_tpch-q19-default.txt-Plan_/plan.txt"
         }
     ],
     "test.test[tpch-q19-default.txt-Results]": [],
