+++ ydb/library/yql/tests/sql/hybrid_file/part2/canondata/result.json
@@ -1569,9 +1569,9 @@
     ],
     "test.test[key_filter-lambda_with_null_filter--Debug]": [
         {
-            "checksum": "8cd403e2726a1352a9ef3bd3362f4f42",
-            "size": 5024,
-            "uri": "https://{canondata_backend}/1809005/ad7c074711ee8d1675aebabbf8025a2c8bd317d8/resource.tar.gz#test.test_key_filter-lambda_with_null_filter--Debug_/opt.yql_patched"
+            "checksum": "70feaa30772d66bec4fe0139254456a7",
+            "size": 5031,
+            "uri": "https://{canondata_backend}/1599023/0e73c4fa67e9c960ff5312b7132f6c7465a2e8d9/resource.tar.gz#test.test_key_filter-lambda_with_null_filter--Debug_/opt.yql_patched"
         }
     ],
     "test.test[key_filter-lambda_with_null_filter--Plan]": [
