+++ ydb/library/yql/tests/sql/suites/join/do_not_suppres_equijoin_input_sorts.cfg
@@ -1,0 +1,3 @@
+in Input1 sorted_by_kv1_opt.txt
+in Input2 sorted_by_kv1.txt
+providers yt
